backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
cryptography==42.0.8
httpx==0.28.1
pydantic==2.11.3
python-dotenv==1.0.1
